因 Cart.vue 及 AddresseeForm 資料彼此傳輸寫更多複雜程式碼，故刪除此元件，將此元件內容改寫到 Cart.vue 中
